+++ commands/credentials.js
@@ -3,7 +3,7 @@
 const co = require('co')
 const cli = require('heroku-cli-util')
 const sortBy = require('lodash.sortby')
-const printf = require('printf')
+const util = require('../lib/util')
 
 function * run (context, heroku) {
   const fetcher = require('../lib/fetcher')(heroku)
@@ -16,22 +16,6 @@ function * run (context, heroku) {
     let attachments = []
     let credentials = []
 
-    function formatAttachment (attachment) {
-      let attName = cli.color.addon(attachment.name)
-
-      let output = [cli.color.dim('as'), attName]
-      let appInfo = `on ${cli.color.app(attachment.app.name)} app`
-      output.push(cli.color.dim(appInfo))
-
-      return output.join(' ')
-    }
-
-    function renderAttachment (attachment, app, isLast) {
-      let line = isLast ? '└─' : '├─'
-      let attName = formatAttachment(attachment)
-      return printf(' %s %s', cli.color.dim(line), attName)
-    }
-
     function presentCredential (cred) {
       let credAttachments = []
       if (cred !== 'default') {
@@ -39,45 +23,7 @@ function * run (context, heroku) {
       } else {
         credAttachments = attachments.filter(a => a.namespace === null)
       }
-      let isForeignApp = (attOrAddon) => attOrAddon.app.name !== app
-      let atts = sortBy(credAttachments,
-        isForeignApp,
-        'name',
-        'app.name'
-      )
-
-      // render each attachment under the credential
-      let attLines = atts.map(function (attachment, idx) {
-        let isLast = (idx === credAttachments.length - 1)
-        return renderAttachment(attachment, app, isLast)
-      })
-
-      let rotationLines = []
-      let credentialStore = credentials.filter(a => a.name === cred)[0]
-      if (credentialStore.state === 'rotating') {
-        let formatted = credentialStore.credentials.map(function (credential, idx) {
-          return {
-            'user': credential.user,
-            'state': credential.state,
-            'connections': credential.connections
-          }
-        })
-        let connectionInformationAvailable = formatted.some((c) => c.connections)
-        if (connectionInformationAvailable) {
-          let prefix = '       '
-          rotationLines.push(`${prefix}Usernames currently active for this credential:`)
-          cli.table(formatted, {
-            printHeader: false,
-            printLine: function (line) { rotationLines.push(line) },
-            columns: [
-              {key: 'user', format: (v, r) => `${prefix}${v}`},
-              {key: 'state', format: (v, r) => (v === 'revoking') ? 'waiting for no connections to be revoked' : v},
-              {key: 'connections', format: (v, r) => `${v} connections`}
-            ]
-          })
-        }
-      }
-      return [cred].concat(attLines).concat(rotationLines).join('\n')
+      return util.presentCredentialAttachments(app, credAttachments, credentials, cred)
     }
 
     try {
